requirements.txt
python-telegram-bot
sentence-transformers
faiss-cpu
requests
python-dotenv
asyncio
openai
sentence-transformers